fix: use correct credentials when invoked multiple times
Since we can't easilly infer which file in `~/.aws/cli/cache` is the one
that correspond to the profile we're trying to log in, we were relying
on the fact that the last modified file should be the one with the
credentials that are relevant to our log in workflow.

Along the way, we discovered that when the cache file is still valid,
`aws sso login` does not modify it.

That means that in some edge case (see hereafter), we could pick the wrong
cache file and modify the `~/.aws/credentials` with valid credentials
that are not matching the correct `[profile-name]`.
That was a nasty bug and it was tricky to troubleshoot.

Our current workaround is a bit dirty: we're systematically deleting the
`~/.aws/cli/cache/*.json` files before running the original `aws sso login`
command.
That guarantees us that the only cache file that would be present is the
one that has been generated during the current log in workflow.

It is certainly sub-optimal but it works for our current usage.