BREAKING CHANGE: adding support of '--profile' arg
In an effort to be a drop-in replacement for the original `aws sso login`
command, we're now supporting the `--profile` arg instead of simply taking
the first argument.

This is a BREAKING CHANGE. See examples.

```
aws-sso-login my-profile-name

aws-sso-login --profile my-profile-name
```

That would ease the adoption for users that were previously using the
original `aws sso login` command.

Note: `AWS_SSO_DEFAULT_PROFILE` and `AWS_PROFILE` environment are still
supported.